pyproject.toml
[project]
name = "bittensor-opinion-bot"
requires-python = "==3.11.*"
version = "0"
dependencies = [
    "Django~=4.2.4",
    "django-csp==3.7",
    "django-constance[database]==3.1.0",
    "django-cors-headers~=4.2.0",
    "django-environ~=0.11.2",
    "django-extensions==3.2.3",
    "django-probes==1.7.0",
    "django-debug-toolbar==4.1.0",
    "django-structlog[celery]==8.0.0",
    "djangorestframework~=3.16.0",
    "drf-spectacular[sidecar]~=0.28.0",
    "celery~=5.3.1",
    "flower~=2.0.0",
    "gunicorn==20.1.0",
    "psycopg[binary]~=3.1.19",
    "redis~=4.6.0",
    "sentry-sdk==1.3.0",
    "ipython~=8.14.0",
    "nox==2024.10.9",
    "more-itertools~=10.3.0",
    "psutil>=5.9.8",
    "prometheus-client~=0.17.0",
    "django-prometheus==2.3.1",
    "django-business-metrics>=1.0.1,<2",
    "channels[daphne]~=4.0",
    "channels-redis~=4.2.0",
    "uvicorn[standard]==0.29",
    "pydantic~=2.0",
    "discord-py~=2.4",
    "httpx>=0.28.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[dependency-groups]
test = [
    'pytest',
    'pytest-django',
    'pytest-xdist',
    'pytest-asyncio',
    'ipdb',
    'freezegun',
]
lint = [
    "ruff",
    "codespell[toml]",
]
type_check = [
    "django-stubs[compatible-mypy]",
    "djangorestframework-stubs[compatible-mypy]",
    "mypy",
    "types-freezegun",
    "types-python-dateutil",
    "types-requests",
]

[tool.ruff]
src = ["app/src"]
line-length = 120

[tool.ruff.lint]
# TODO add D
select = [
    "E", "F", "I", "UP", "S",
    "TC005",
]
# TODO: remove E501 once docstrings are formatted
ignore = [
    "D100", "D105", "D107", "D200", "D202", "D203", "D205", "D212", "D400", "D401", "D415",
    "D101", "D102","D103", "D104", # TODO remove once we have docstring for all public methods
    "E501", # TODO: remove E501 once docstrings are formatted
    "S101", "S108", "S603", "S607",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/test*/**.py" = ["D", "F403", "F405", "S106", "S311"]

[tool.codespell]
skip = '*.min.js,*.lock,*/monitoring_certs/*'
ignore-words-list = 'datas'
